// src/types/hotelDetails.ts
export interface ListingDetail {
  id?: number;
  user_id?: number;
  ownerdetails?: Ownerdetails;
  name?: string;
  name_hindi?: string;
  rate_type?: null | string;
  description?: string;
  description_hindi?: string;
  order?: number;
  property_type?: string;
  property_chain?: string;
  hotel_chain?: number;
  hotel_chain_name?: string;
  star_category?: string | number;
  no_of_rooms?: number;
  lat?: number;
  lon?: number;
  no_of_floors?: number;
  no_of_buildings?: number;
  is_favourite?: number;
  images?: ListingImage[];
  documents_required_for_check_in?: boolean;
  contactdetails?: Contactdetail[];
  address?: string;
  city?: number;
  state?: number;
  country?: number;
  guest_cancellation_type?: null | string;
  check_in?: string;
  check_out?: string;
  check_in_age?: string;
  check_in_policies?: string;
  step_completed?: number;
  status?: boolean;
  status_remark?: string;
  instant_booking?: boolean;
  category?: string;
  b2c_commission_rate?: number;
  website?: string;
  meta_title?: string;
  meta_keywords?: string;
  meta_description?: string;
  meta_tags?: string;
  created?: string | Date;
  video_request_status?: null | string;
  check_in_document?: string;
  servicedetails?: Servicedetail[];
  b2b_commission_rate?: number;
  videos?: any[];
  city_name?: string;
  state_name?: string;
  country_short_name?: string;
  country_name?: string;
  show_landing?: boolean;
  show_landing2?: boolean;
  show_featured?: boolean;
  show_popular?: boolean;
  show_top_rated?: boolean;
  show_best_deals?: boolean;
  review_rating?: number;
  sbr_rate?: number;
  review_rating_count?: number;
  video_url?: null | string;
  rooms?: HotelRoom[];
  facilities?: number[];
  facilitiesDetails?: FacilitiesDetail[];
  other_rules?: string;
  highlights?: string;
  guest_description?: null | string;
  title_highlights?: string;
  meta_names?: string;
  property_highlights?: any[];
  property_highlights_details?: any[];
  cancellation_policies?: CancellationPolicy[];
  taxation_details?: TaxationDetail[];
  deduction_details?: any[];
  promotion_details?: PromotionDetails;
}

export interface CancellationPolicy {
  id?: number;
  cancellation_policy?: number;
  policy_name?: string;
  policy_no_of_days?: number;
  policy_description?: string;
  policy_start_date?: string;
  policy_end_date?: string;
  created?: string | Date;
  status?: boolean;
  status_remark?: null | string;
  created_by?: number;
}

export interface Contactdetail {
  id?: number;
  name?: string;
  email?: string;
  mobile?: string;
  contact_type?: string;
}

export interface FacilitiesDetail {
  id?: number;
  name?: string;
  name_hindi?: string;
  parent?: null | number;
  amenities_tags?: RoomTypeDetails[];
  image?: string;
  created?: string | Date;
}

export interface RoomTypeDetails {
  id?: number;
  name?: string;
  created?: string | Date;
  created_by?: number;
}

export interface ListingImage {
  id?: number;
  file?: string;
  cover_photo?: boolean;
  images_tag?: any[];
  order?: number;
  status?: number;
  status_remark?: null | string;
  room_images_tag?: any[];
}

export interface Ownerdetails {
  url?: string;
  id?: number;
  username?: string;
  first_name?: string;
  last_name?: string;
  email?: string;
  user_type?: string;
  company_name?: null | string;
  city?: string;
  full_name?: string;
}

export interface PromotionDetails {
  basic_promotion?: any[];
  early_bird_promotion?: any[];
  last_minute_promotion?: any[];
  long_stay_promotion?: any[];
  advance_purchase_promotion?: any[];
  hotel_offer_promotion?: any[];
  member_only_promotion?: any[];
  value_added_promotion?: any[];
}

export interface HotelRoom {
  room_name?: string;
  created?: string | Date;
  description?: string;
  suitable_for_kids?: boolean;
  pets_allowed?: boolean;
  smoking_allowed?: boolean;
  room_type?: number;
  id?: number;
  room_location?: string;
  room_view?: string;
  no_of_rooms?: string | number;
  costume_room_name?: string;
  base_adults?: number;
  maximum_adults?: number;
  maximum_children?: number;
  dimensions?: string;
  bed_type?: string;
  images?: ListingImage[];
  no_of_beds?: number;
  extra_bed_allowed?: boolean;
  extra_bed_type?: string;
  maximum_occupancy?: number;
  ep_highlights?: string;
  cp_highlights?: string;
  map_highlights?: string;
  ap_highlights?: string;
  room_type_details?: RoomTypeDetails;
  videos?: any[];
  facilitiesDetails?: FacilitiesDetail[];
  facilities?: number[];
  ap_rate?: number;
  map_rate?: number;
  cp_rate?: number;
  ep_rate?: number;
  plans?: RoomPlan[];
}

export interface RoomPlan {
  id?: number;
  plan?: number;
  plan_name?: string;
  plan_items?: RoomTypeDetails[];
  room?: number;
}

export interface Servicedetail {
  id?: number;
  name?: string;
  serviceType?: string;
  openTime?: string;
  closeTime?: string;
  typeofService?: string;
  images?: any[];
}

export interface TaxationDetail {
  id?: number;
  country?: number;
  country_name?: string;
  name?: string;
  code?: string;
  tax_type?: string;
  amount_or_percentage?: number;
}